Add files via upload
You are given a hexagonal grid consisting of two rows, each row consisting of  cells. The cells of the first row are labelled  and the cells of the second row are labelled .

For example, for :

Grid Shape

(Note that the  is connected with .)

Your task is to tile this grid with  tiles that look like the following:

Orientations

As you can see above, there are three possible orientations in which a tile can be placed.

Your goal is to tile the whole grid such that every cell is covered by a tile, and no two tiles occupy the same cell. To add to the woes, certain cells of the hexagonal grid are blackened. No tile must occupy a blackened cell.

Is it possible to tile the grid?

Here's an example. Suppose we want to tile this grid:

Example Blank

Then we can do the tiling as follows:

Example Tiled